==> 3. SII value output creation (rotavirus vaccination).Rmd <==
---
title: '**Rotavirus vaccine uptake by deprivation in England (regression and SII outputs)**'
author: "Aidan Flatt; Dan Hungerford; David Taylor-Robinson"
output:
  html_document:
    df_print: paged
  pdf_document: default
editor_options:
  chunk_output_type: inline
---

```{r echo=FALSE, error=FALSE, message=FALSE, warning=FALSE, results='hide'}
rm(list=ls())

##load packages
library(readr)
library(devtools)
library(data.table) 
library(lme4)
library(sp)
library(stringr)
library(dplyr)
library(knitr)     
library(kableExtra) 
library(tidyverse)
library(ggplot2)
library(plotly)

```

```{r echo=FALSE, error=FALSE, message=FALSE, warning=FALSE, results='hide'}

#read COVER quarterly vaccine data combined
# original source of data https://www.gov.uk/government/collections/vaccine-uptake#cover-of-vaccination-evaluated-rapidly-programme - see previous R code files in repository for data cleaning and preparation for analysis
data <- read_csv("GP rota merge length ways.csv")
data <- data[,c(2,3,4,5,6,7)]
names(data)[2] <- "Practice_code"

#Read in deprivation data
# original source of data Office for Health Improvement & Disparities.  https://fingertips.phe.org.uk/search/deprivation see previous R code files in repository for data cleaning and preparation for analysis
deprivation_data <- read_csv("GP codes with deprivation.csv")
names(deprivation_data)[2] <- "Practice_code"
names(deprivation_data)[3] <- "Practice_name"
names(deprivation_data)[5] <- "IMD_2019_score"
names(deprivation_data)[6] <- "IMD_decile"
deprivation_data <- deprivation_data[,c(2,3,5,6)]

merged_data <- merge(data, deprivation_data, by = "Practice_code", all.x = TRUE)

##Rename the date windows to quarter numbers 1 through 16
merged_data <- merged_data %>%
  mutate('Quarter_number' = case_when(
    Date_quarter == "Apr19Jun19" ~ "01",
    Date_quarter == "Jul19Sep19" ~ "02",
    Date_quarter == "Oct19Dec19" ~ "03",
    Date_quarter == "Jan20Mar20" ~ "04",
    Date_quarter == "Apr20Jun20" ~ "05",
    Date_quarter == "Jul20Sep20" ~ "06",
    Date_quarter == "Oct20Dec20" ~ "07",
    Date_quarter == "Jan21Mar21" ~ "08",
    Date_quarter == "Apr21Jun21" ~ "09",
    Date_quarter == "Jul21Sep21" ~ "10",
    Date_quarter == "Oct21Dec21" ~ "11",
    Date_quarter == "Jan22Mar22" ~ "12",
    Date_quarter == "Apr22Jun22" ~ "13",
    Date_quarter == "Jul22Sep22" ~ "14",
    Date_quarter == "Oct22Dec22" ~ "15",
    Date_quarter == "Jan23Mar23" ~ "16",
    TRUE ~ Date_quarter
  ))

class(merged_data$`Quarter_number`)
merged_data$`Quarter_number` <- as.numeric(merged_data$`Quarter_number`)

## Make a new column with numerators to allow to undertake weighted mean:
merged_data$Vacc_numerator <- merged_data$`12m Denom` * (merged_data$`Rotavacc% 12m` /100)

## Remove excess row
merged_data<-subset(merged_data, Practice_code!=0)


## Unknown LA code and no assigned GP
merged_data<-subset(merged_data, !is.na(`ODS Upper Tier LA Code`))


## Remove GP pracs with missing deprivation data
merged_data<-subset(merged_data,!is.na(IMD_decile))


## Remove specific LAs (Isle of Scilly 906, City of London 714)
merged_data <- merged_data %>%
  filter(!`ODS Upper Tier LA Code` %in% c(906, 714))

##Additionally for rotavirus remove Surrey 805 and Bradford 209, removed due to a data derived anomaly
##(This step can be left out for other vaccines)

merged_data <- merged_data %>%
  filter(!`ODS Upper Tier LA Code` %in% c(209, 805))

```

#Creation of a new df where the weighted_deprivation_rank is created to assess the SII. GP practices are first ranked in order of IMD for each quarter. The total population of children in each GP practice is then calculated as a proportion of the total number of children in all GP practices for each quarter. Following this the cumulative number of proportion of children in GP practices of increasing order of deprivation is calculated (ie the least deprived is added to the next least deprived etc etc until reaching the most deprived practice). The mid-point of the range between each GP practice is then calculated and called the 'weighted_deprivation_rank'. 0 represents the least deprived practice, and 1 represents the most deprived practice.

```{r}
# Assuming merged_data is the data frame containing data for all quarters (Q1 to Q16)

# Define a list to store the modified data frames
df_list <- list()

# Create a loop through quarters Q1 to Q16
for (quarter_number in 1:16) {
  # Filter data for the current quarter (Q1, Q2, ..., Q16)
  dfQ <- merged_data[merged_data$Quarter_number == quarter_number, ]
  
  # Create a column ranking by IMD score
  order.scores <- order(dfQ$IMD_2019_score)
  dfQ$rank <- NA
  dfQ$rank[order.scores] <- 1:nrow(dfQ)
  
  # Calculate the sum of the population ie '12m Denom' and exclude NA values
  total_denominator <- sum(dfQ$`12m Denom`, na.rm = TRUE)
  dfQ <- dfQ %>% mutate(percpop = `12m Denom` / total_denominator) %>%
    filter(!is.na(percpop))
  
  # Re-order by rank, and calculate the cumulative proportion of the population in each rank
  dfQ <- dfQ[order(dfQ$rank),]
  dfQ <- dfQ %>%  mutate(cumul_pop = cumsum(percpop))
  
  # Find the mid-point of the range of values for ranks
  dfQ <- dfQ %>% mutate(weighted_deprivation_rank = (dplyr::lag(cumul_pop) + (percpop / 2)))
  
  # Rank 1 is NA as is used as a lag function, so we need to find a value for this
  dfQ$cumul_pop_halved <- (dfQ$cumul_pop / 2)
  dfQ$weighted_deprivation_rank[is.na(dfQ$weighted_deprivation_rank)] <- as.character(dfQ$cumul_pop_halved[is.na(dfQ$weighted_deprivation_rank)])
  dfQ$weighted_deprivation_rank <- as.numeric(dfQ$weighted_deprivation_rank)
  
  # Add the modified data frame for the quarter being analysed to the list (and repeat for each of the 16 quarters)
  df_list[[quarter_number]] <- dfQ
}
#End of loop


# Print or access the modified data frames for each quarter from df_list
# For example, df_list[[1]] will be the modified data frame for Q1, df_list[[2]] for Q2, and so on.

merged_df <- do.call(rbind, df_list)

#Re-write quarter numbers so first time point is 0 instead of 1

merged_df <- merged_df %>%
  mutate('Quarter_number' = case_when(
    Date_quarter == "Apr19Jun19" ~ "00",
    Date_quarter == "Jul19Sep19" ~ "01",
    Date_quarter == "Oct19Dec19" ~ "02",
    Date_quarter == "Jan20Mar20" ~ "03",
    Date_quarter == "Apr20Jun20" ~ "04",
    Date_quarter == "Jul20Sep20" ~ "05",
    Date_quarter == "Oct20Dec20" ~ "06",
    Date_quarter == "Jan21Mar21" ~ "07",
    Date_quarter == "Apr21Jun21" ~ "08",
    Date_quarter == "Jul21Sep21" ~ "09",
    Date_quarter == "Oct21Dec21" ~ "10",
    Date_quarter == "Jan22Mar22" ~ "11",
    Date_quarter == "Apr22Jun22" ~ "12",
    Date_quarter == "Jul22Sep22" ~ "13",
    Date_quarter == "Oct22Dec22" ~ "14",
    Date_quarter == "Jan23Mar23" ~ "15",
    TRUE ~ Date_quarter
  ))

class(merged_df$`Quarter_number`)
merged_df$`Quarter_number` <- as.numeric(merged_df$`Quarter_number`)

```

#Creation of a spaghetti plot for a random sample of 600 GP pracs (approx 10% of total pracs) to look at potential outliers in uptake
```{r}

# Set the number of GP practices you want to include in the random sample
sample_size600 <- 600  # You can adjust this to your desired sample size

# Step 1: Generate a random sample of GP practices (GP_code)
set.seed(123)  # Set a seed for reproducibility
random600_gp_practices <- sample(unique(merged_df$Practice_code), sample_size600, replace = FALSE)

# Step 2: Filter the data to include only the rows corresponding to the random GP practices
filtered_data600 <- merged_df[merged_df$Practice_code %in% random600_gp_practices, ]

ggplot_object <- ggplot(data = filtered_data600, aes(x = Quarter_number, y = `Rotavacc% 12m`, group = Practice_code, color = Practice_code)) +
  geom_line() +
  xlab("Quarter Number") +
  ylab("Rotavacc% 12m") +
  ggtitle("Spaghetti Plot of Rotavacc% 12m over Quarter Number for random sample of 600 GP Practices in England") +
  theme_minimal() +
  theme(legend.position = "bottom")

# Step 4: Convert ggplot object to plotly object
plotly_object <- ggplotly(ggplot_object)

# Print the interactive plot
plotly_object


```

#Fit the the random effect models for the regression analysis
```{r}

#weighted_deprivation_rank value is now used as a continuous exposure variable in the regression model

#Random effects for LA code
names(merged_df)[2] <- "LA_code"

#Fit the restricted model
mod1 <- lmer(`Rotavacc% 12m` ~ Quarter_number * weighted_deprivation_rank + (1 | LA_code), data = merged_df, control = lmerControl(optimizer = "bobyqa"))
summary(mod1)

mod1a <- lmer(`Rotavacc% 12m` ~ Quarter_number * weighted_deprivation_rank + (1 | Practice_code), data = merged_df, control = lmerControl(optimizer = "bobyqa"))
summary(mod1a)

#Fit the extended model
mod2 <- lmer(`Rotavacc% 12m` ~ Quarter_number * weighted_deprivation_rank + (Quarter_number | LA_code), data = merged_df, control = lmerControl(optimizer = "bobyqa"))
summary(mod2)

#The models have different fixed and random effect specifications. We are testing whether adding random slopes for Quarter_number significantly improves the model fit compared to the model with only random intercepts.

#Perform ANOVA between the two models
anova_result <- anova(mod1, mod2)
print(anova_result)

```

#Using model 2, create a table summarising the SII for the whole time period (16 quarters / 4 years)
```{r}
mod_whole_time <- mod2
SII_whole <- cbind(fixef(mod_whole_time), confint(mod_whole_time, "beta_", method="Wald"))
print(SII_whole)
row.names(SII_whole)[3] <- "rota_SII_whole_time"

kable(SII_whole, format = "html")
rota_SII_whole_time_table <- kable(SII_whole, format = "html")

```

#Create time points across the 16 quarters / 4 years to allow SII outputs for different time periods around the covid-19 pandemic
```{r echo=FALSE, error=FALSE, message=FALSE, warning=FALSE, results='hide'}

#Calculate SIIs for various time points / periods (Q0/start, pre-pandemic, pandemic, post-pandemic, Q15/end)

dfQ0 <- merged_df %>%
  filter(Quarter_number == "0")
modQ0 <- lmer(`Rotavacc% 12m` ~ weighted_deprivation_rank + (1 | LA_code), data = dfQ0, control = lmerControl(optimizer = "bobyqa"))
summary(modQ0)
SII_Q0 <- cbind(fixef(modQ0), confint(modQ0, "beta_", method="Wald"))
row.names(SII_Q0)[2] <- "rota_SII_Q0"
SII_Q0 <- SII_Q0[2, , drop = FALSE]

prepan_quarters <- c("0", "1", "2", "3")
df_prepan <- merged_df %>%
  filter(Quarter_number %in% prepan_quarters)
mod_prepan <- lmer(`Rotavacc% 12m` ~ weighted_deprivation_rank + (Quarter_number | LA_code), data = df_prepan, control = lmerControl(optimizer = "bobyqa"))
summary(mod_prepan)
SII_prepan <- cbind(fixef(mod_prepan), confint(mod_prepan, "beta_", method="Wald"))
row.names(SII_prepan)[2] <- "rota_SII_pre_pandemic"
SII_prepan <- SII_prepan[2, , drop = FALSE]

pan_quarters <- c("4", "5", "6", "7", "8", "9", "10", "11")
df_pan <- merged_df %>%
  filter(Quarter_number %in% pan_quarters)
mod_pan <- lmer(`Rotavacc% 12m` ~ weighted_deprivation_rank + (Quarter_number | LA_code), data = df_pan, control = lmerControl(optimizer = "bobyqa"))
summary(mod_pan)
SII_pan <- cbind(fixef(mod_pan), confint(mod_pan, "beta_", method="Wald"))
row.names(SII_pan)[2] <- "rota_SII_pandemic"
SII_pan <- SII_pan[2, , drop = FALSE]

postpan_quarters <- c("12", "13", "14", "15")
df_postpan <- merged_df %>%
  filter(Quarter_number %in% postpan_quarters)
mod_postpan <- lmer(`Rotavacc% 12m` ~ weighted_deprivation_rank + (Quarter_number | LA_code), data = df_postpan, control = lmerControl(optimizer = "bobyqa"))
summary(mod_postpan)
SII_postpan <- cbind(fixef(mod_postpan), confint(mod_postpan, "beta_", method="Wald"))
row.names(SII_postpan)[2] <- "rota_SII_post_pandemic"
SII_postpan <- SII_postpan[2, , drop = FALSE]

dfQ15 <- merged_df %>%
  filter(Quarter_number == "15")
modQ15 <- lmer(`Rotavacc% 12m` ~ weighted_deprivation_rank + (1 | LA_code), data = dfQ15, control = lmerControl(optimizer = "bobyqa"))
summary(modQ15)
SII_Q15 <- cbind(fixef(modQ15), confint(modQ15, "beta_", method="Wald"))
row.names(SII_Q15)[2] <- "rota_SII_Q15"
SII_Q15 <- SII_Q15[2, , drop = FALSE]

SII_time_periods <- rbind(SII_Q0, SII_prepan, SII_pan, SII_postpan, SII_Q15)


```

##Create a table summarising these SIIs for different time periods around the pandemic (Q0/start, pre-pandemic, pandemic, post-pandemic, Q15/end)
```{r}
kable(SII_time_periods, format = "html")
rota_SII_time_periods_table <- kable(SII_time_periods, format = "html")

```

###Create SIIs values for annual trend assessment (ie for each of the 4 years included in the analysis)
```{r echo=FALSE, error=FALSE, message=FALSE, warning=FALSE, results='hide'}

time2019_2020 <- c("0", "1", "2", "3")
df_2019_2020 <- merged_df %>%
  filter(Quarter_number %in% time2019_2020)
mod1920 <- lmer(`Rotavacc% 12m` ~ weighted_deprivation_rank + (Quarter_number | LA_code), data = df_2019_2020, control = lmerControl(optimizer = "bobyqa"))
summary(mod1920)
SII_2019_2020 <- cbind(fixef(mod1920), confint(mod1920, "beta_", method="Wald"))
row.names(SII_2019_2020)[2] <- "rota_SII_2019_2020"
SII_2019_2020 <- SII_2019_2020[2, , drop = FALSE]

time2020_2021 <- c("4", "5", "6", "7")
df_2020_2021 <- merged_df %>%
  filter(Quarter_number %in% time2020_2021)
mod2021 <- lmer(`Rotavacc% 12m` ~ weighted_deprivation_rank + (Quarter_number | LA_code), data = df_2020_2021, control = lmerControl(optimizer = "bobyqa"))
summary(mod2021)
SII_2020_2021 <- cbind(fixef(mod2021), confint(mod2021, "beta_", method="Wald"))
row.names(SII_2020_2021)[2] <- "rota_SII_2020_2021"
SII_2020_2021 <- SII_2020_2021[2, , drop = FALSE]

time2021_2022 <- c("8", "9", "10", "11")
df_2021_2022 <- merged_df %>%
  filter(Quarter_number %in% time2021_2022)
mod2122 <- lmer(`Rotavacc% 12m` ~ weighted_deprivation_rank + (Quarter_number | LA_code), data = df_2021_2022, control = lmerControl(optimizer = "bobyqa"))
summary(mod2122)
SII_2021_2022 <- cbind(fixef(mod2122), confint(mod2122, "beta_", method="Wald"))
row.names(SII_2021_2022)[2] <- "rota_SII_2021_2022"
SII_2021_2022 <- SII_2021_2022[2, , drop = FALSE]

time2022_2023 <- c("12", "13", "14", "15")
df_2022_2023 <- merged_df %>%
  filter(Quarter_number %in% time2022_2023)
mod2223 <- lmer(`Rotavacc% 12m` ~ weighted_deprivation_rank + (Quarter_number | LA_code), data = df_2022_2023, control = lmerControl(optimizer = "bobyqa"))
summary(mod2223)
SII_2022_2023 <- cbind(fixef(mod2223), confint(mod2223, "beta_", method="Wald"))
row.names(SII_2022_2023)[2] <- "rota_SII_2022_2023"
SII_2022_2023 <- SII_2022_2023[2, , drop = FALSE]

SII_annual_summary <- rbind(SII_2019_2020, SII_2020_2021, SII_2021_2022, SII_2022_2023)
```

##Create a table summarising these SIIs for annual trends
```{r}
kable(SII_annual_summary, format = "html")
rota_SII_annual_summary_table <- kable(SII_annual_summary, format = "html")

```
